+++ app/controllers/community_tool_reviews_controller.rb
@@ -2,10 +2,10 @@
 
 # Allows user to review community tools.
 class CommunityToolReviewsController < ApplicationController
-  before_action :authenticate_user!, except: [:index]
+  before_action :authenticate_user!, except: [:index, :show]
   before_action :find_community_tool_or_redirect
   before_action :find_existing_community_tool_review_and_redirect_to_edit, only: [:new, :create]
-  before_action :find_community_tool_review_or_redirect, only: [:show, :edit, :update, :destroy]
+  before_action :find_community_tool_review_or_redirect, only: [:edit, :update, :destroy]
 
   # GET /community_tools/1/reviews
   def index
@@ -14,6 +14,8 @@ class CommunityToolReviewsController < ApplicationController
 
   # GET /community_tools/1/reviews/1
   def show
+    @community_tool_review = @community_tool.community_tool_reviews.find_by_id(params[:id])
+    redirect_to community_show_tool_path(@community_tool) unless @community_tool_review
   end
 
   # GET /community_tools/1/reviews/new
@@ -29,6 +31,7 @@ class CommunityToolReviewsController < ApplicationController
   def create
     @community_tool_review = @community_tool.community_tool_reviews.where(user_id: current_user.id).new(community_tool_review_params)
     if @community_tool_review.save
+      @community_tool_review.create_notification!
       redirect_to community_show_tool_path(@community_tool), notice: 'Review was successfully created.'
     else
       render :new
@@ -38,6 +41,7 @@ class CommunityToolReviewsController < ApplicationController
   # PATCH /community_tools/1/reviews/1
   def update
     if @community_tool_review.update(community_tool_review_params)
+      @community_tool_review.create_notification!
       redirect_to community_show_tool_path(@community_tool), notice: 'Review was successfully updated.'
     else
       render :edit
@@ -53,7 +57,7 @@ class CommunityToolReviewsController < ApplicationController
   private
 
   def find_community_tool_or_redirect
-    @community_tool = CommunityTool.current.find_by(id: params[:community_tool_id])
+    @community_tool = CommunityTool.current.find_by_param(params[:community_tool_id])
     redirect_without_community_tool
   end