+++ app/models/community_tool_review.rb
@@ -10,6 +10,19 @@ class CommunityToolReview < ApplicationRecord
   # Model Relationships
   belongs_to :community_tool, touch: true
   belongs_to :user
+  has_many :notifications
 
   # Model Methods
+
+  def create_notification!
+    notification = community_tool.user.notifications
+                                 .where(community_tool_id: community_tool_id, community_tool_review_id: id)
+                                 .first_or_create
+    notification.mark_as_unread!
+  end
+
+  def destroy
+    notifications.destroy_all
+    super
+  end
 end