+++ db/migrate/20161214200743_add_slug_to_community_tools.rb
@@ -1,0 +1,6 @@
+class AddSlugToCommunityTools < ActiveRecord::Migration[5.0]
+  def change
+    add_column :community_tools, :slug, :string
+    add_index :community_tools, :slug, unique: true
+  end
+end